// src/services/generated/apiTypes/relationship.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type Relationship = typeof Relationship[keyof typeof Relationship];


export const Relationship = {
  padre: 'padre',
  madre: 'madre',
  hijo: 'hijo',
  abuelo: 'abuelo',
  nieto: 'nieto',
  tutor: 'tutor',
  otro: 'otro',
} as const;