// src/services/generated/apiTypes/religion.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Laravel
 * OpenAPI spec version: 0.0.1
 */

export type Religion = typeof Religion[keyof typeof Religion];


export const Religion = {
  catholic: 'catholic',
  christian: 'christian',
  jehovahs_witness: 'jehovahs_witness',
  mormon: 'mormon',
  none: 'none',
  other: 'other',
} as const;